+++ NEWS.md
@@ -1,5 +1,7 @@
 # httpuv (development version)
 
+* Make `getStatusDescription` available from R. (#321)
+
 # httpuv 1.6.12
 
 * New `runStaticServer()` provides a convenient interface for serving a directory of static files. (#380)